+++ app/src/test/java/com/windsoft/se/project/GameQuestionTest.java
@@ -4,15 +4,14 @@ package com.windsoft.se.project;
  * Created by João Lucas on 10/12/2017.
  */
 
-import com.windsoft.se.project.questao.LEVEL;
-import com.windsoft.se.project.questao.Question;
+import com.windsoft.se.project.model.GameQuestion;
+import com.windsoft.se.project.model.LEVEL;
+import com.windsoft.se.project.model.Question;
 
-import org.junit.Assert;
 import org.junit.Test;
 
 import java.util.HashSet;
-import java.util.Set;
-import java.util.regex.Pattern;
+
 import static org.junit.Assert.*;
 
 public class GameQuestionTest {
